+++ src/app/(main)/discover/loading.tsx
@@ -1,6 +1,6 @@
 import { Center } from 'react-layout-kit';
 
-import CircleLoading from '@/components/CircleLoading';
+import CircleLoading from '@/components/Loading/BrandTextLoading';
 
 export default () => (
   <Center height={'90vh'} width={'100%'}>
