+++ src/components/Loading/BrandTextLoading/LobeChatText/SVG.tsx
