+++ src/components/Loading/BrandTextLoading/index.tsx
@@ -1,11 +1,11 @@
 import { Center } from 'react-layout-kit';
 
-import LobeChatTextLoading from '@/components/LobeChatTextLoading';
+import LobeChatText from './LobeChatText';
 
 export default () => {
   return (
     <Center height={'100%'} width={'100%'}>
-      <LobeChatTextLoading />
+      <LobeChatText />
     </Center>
   );
 };
