+++ src/features/Setting/SettingContainer.tsx
@@ -34,6 +34,7 @@ const SettingContainer = memo<PropsWithChildren<SettingContainerProps>>(
         {addonBefore}
         <Flexbox
           gap={64}
+          flex={1}
           paddingInline={mobile ? undefined : 24}
           style={{
             maxWidth,
